Replace publisherlistenerasync by dyn version
