+++ dds/src/dds_async/publisher_listener.rs
@@ -1,4 +1,4 @@
-use std::future::Future;
+use std::{future::Future, pin::Pin};
 
 use crate::{
     infrastructure::status::{
@@ -11,38 +11,50 @@ use crate::{
 /// This trait represents a listener object which can be associated with the [`PublisherAsync`](super::publisher::Publisher) entity.
 pub trait PublisherListenerAsync {
     /// Method that is called when any writer belonging to this publisher reports a liveliness lost status.
-    fn on_liveliness_lost(
-        &mut self,
-        _the_writer: &dyn AnyDataWriter,
-        _status: LivelinessLostStatus,
-    ) -> impl Future<Output = ()> + Send {
-        std::future::ready(())
+    fn on_liveliness_lost<'a, 'b>(
+        &'a mut self,
+        the_writer: &'b (dyn AnyDataWriter + Sync),
+        status: LivelinessLostStatus,
+    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'b>>
+    where
+        'a: 'b,
+    {
+        Box::pin(std::future::ready(()))
     }
 
     /// Method that is called when any writer belonging to this publisher reports an offered deadline missed status.
-    fn on_offered_deadline_missed(
-        &mut self,
-        _the_writer: &dyn AnyDataWriter,
-        _status: OfferedDeadlineMissedStatus,
-    ) -> impl Future<Output = ()> + Send {
-        std::future::ready(())
+    fn on_offered_deadline_missed<'a, 'b>(
+        &'a mut self,
+        the_writer: &'b (dyn AnyDataWriter + Sync),
+        status: OfferedDeadlineMissedStatus,
+    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'b>>
+    where
+        'a: 'b,
+    {
+        Box::pin(std::future::ready(()))
     }
 
     /// Method that is called when any writer belonging to this publisher reports an offered incompatible qos status.
-    fn on_offered_incompatible_qos(
-        &mut self,
-        _the_writer: &dyn AnyDataWriter,
-        _status: OfferedIncompatibleQosStatus,
-    ) -> impl Future<Output = ()> + Send {
-        std::future::ready(())
+    fn on_offered_incompatible_qos<'a, 'b>(
+        &'a mut self,
+        the_writer: &'b (dyn AnyDataWriter + Sync),
+        status: OfferedIncompatibleQosStatus,
+    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'b>>
+    where
+        'a: 'b,
+    {
+        Box::pin(std::future::ready(()))
     }
 
     /// Method that is called when any writer belonging to this publisher reports a publication matched status.
-    fn on_publication_matched(
-        &mut self,
-        _the_writer: &dyn AnyDataWriter,
-        _status: PublicationMatchedStatus,
-    ) -> impl Future<Output = ()> + Send {
-        std::future::ready(())
+    fn on_publication_matched<'a, 'b>(
+        &'a mut self,
+        the_writer: &'b (dyn AnyDataWriter + Sync),
+        status: PublicationMatchedStatus,
+    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'b>>
+    where
+        'a: 'b,
+    {
+        Box::pin(std::future::ready(()))
     }
 }
