+++ dds/src/implementation/utils/sync_listener.rs
@@ -154,40 +154,52 @@ impl<T> PublisherListenerAsync for ListenerSyncToAsync<T>
 where
     T: PublisherListener,
 {
-    fn on_liveliness_lost(
-        &mut self,
-        the_writer: &dyn AnyDataWriter,
+    fn on_liveliness_lost<'a, 'b>(
+        &'a mut self,
+        the_writer: &'b (dyn AnyDataWriter + Sync),
         status: LivelinessLostStatus,
-    ) -> impl Future<Output = ()> + Send {
+    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'b>>
+    where
+        'a: 'b,
+    {
         tokio::task::block_in_place(|| self.0.on_liveliness_lost(the_writer, status));
-        std::future::ready(())
+        Box::pin(std::future::ready(()))
     }
 
-    fn on_offered_deadline_missed(
-        &mut self,
-        the_writer: &dyn AnyDataWriter,
+    fn on_offered_deadline_missed<'a, 'b>(
+        &'a mut self,
+        the_writer: &'b (dyn AnyDataWriter + Sync),
         status: OfferedDeadlineMissedStatus,
-    ) -> impl Future<Output = ()> + Send {
+    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'b>>
+    where
+        'a: 'b,
+    {
         tokio::task::block_in_place(|| self.0.on_offered_deadline_missed(the_writer, status));
-        std::future::ready(())
+        Box::pin(std::future::ready(()))
     }
 
-    fn on_offered_incompatible_qos(
-        &mut self,
-        the_writer: &dyn AnyDataWriter,
+    fn on_offered_incompatible_qos<'a, 'b>(
+        &'a mut self,
+        the_writer: &'b (dyn AnyDataWriter + Sync),
         status: OfferedIncompatibleQosStatus,
-    ) -> impl Future<Output = ()> + Send {
+    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'b>>
+    where
+        'a: 'b,
+    {
         tokio::task::block_in_place(|| self.0.on_offered_incompatible_qos(the_writer, status));
-        std::future::ready(())
+        Box::pin(std::future::ready(()))
     }
 
-    fn on_publication_matched(
-        &mut self,
-        the_writer: &dyn AnyDataWriter,
+    fn on_publication_matched<'a, 'b>(
+        &'a mut self,
+        the_writer: &'b (dyn AnyDataWriter + Sync),
         status: PublicationMatchedStatus,
-    ) -> impl Future<Output = ()> + Send {
+    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'b>>
+    where
+        'a: 'b,
+    {
         tokio::task::block_in_place(|| self.0.on_publication_matched(the_writer, status));
-        std::future::ready(())
+        Box::pin(std::future::ready(()))
     }
 }
